RAG-chatbot-langchain/requirements.txt
langchain 
langchain-community
torch 
transformers[torch] 
sentence-transformers 
datasets 
faiss-cpu 
gradio
python-dotenv